feat: remove PDF Permissions tool and update related descriptions
